First commit!
The very first commit of the MusicManager project.

FormatTools is working very well. NamingTools needs to be completed
before the distribution is technically anywhere near complete.

In the future, a BUGS file would be good to keep around. Right now,
there aren't any because it's pre-alpha and we can't even test it all
the way.