Partial functionality achieved
MusicManager now has code to find broken folders, create appropriately
named directories according to its best guess of file parameters. Python
2.7 has way too many unicode bugs; I wish mutagen existed for 3.3.

BUGS and ROADMAP added; README updated to not suck so hard. All should
be fairly self-explanatory.

fixlist, broken, and proposed are all text file outputs which contain
(in order) a list of directories to fix, a list of directories that
broke during testing, and a preliminary of the proposed fixes.